Bug fixes in .bashrc: Load an older version of ncview
Split off the module load command for ncview and nco from the module
load command for netCDF-4.1.3.  The nco load command will actually
force loading of a more recent module, which is not what we want.

For now, use the centos6/ncview-2.1.2 module, which will not override
the loaded netCDF version.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>