Update .bashrc for PGI compiler and separate netcdf-fortran libraries
Add module load commands for the PGI compiler on Odyssey.

Also now define NETCDF_F_HOME, which is the netCDF-Fortran root
directory.  For the AS cluster, this will default to NETCDF_HOME.

Some versions of netCDF on Odyssey have the netCDF and netCDF-Fortran
libraries installed in separate data paths, thus necessitating these
modifications.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>