Update setup files to simplify transfers between Odyssey and AS
All file transfers to/from Odyssey will need to be done from the AS system.
The config file here needs to be copied to your ~/.ssh/ directory. In that
file, we define a Host called OD with the ControlMaster option set to
"auto." ControlMaster enables sharing of multiple sessions over a single
connection and means that you won't have to enter your password and
verification code for subsequent ssh and scp commands. USERNAME in this
file needs to be changed for each individual user.

The .bashrc file has been updated to include an alias "sshod" that will set
up a ssh connection to Odyssey in the background. Once established, you
can use "scp OD:REMOTEFILE LOCALFILE" to copy files from Odyssey to AS or
"scp LOCALFILE OD:REMOTEFILE" to copy files from AS to Odyssey. A check has
been added to the top of the .bashrc file to determine if an interactive
shell is being used. This will avoid error messages when using the scp
command.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>