Remove test for interactive shell from .bashrc file
The test for an interactive shell was added to avoid an XTERM warning when
using the scp command to copy files to/from Odyssey from AS. This is only
needed if you have a .bashrc file that uses $TERM (e.g. to personalize the
Unix prompt). $XTERM is not used in the default .bashrc file, so this
if statement can be removed.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>