Now set OMP_NUM_THREADS to SLURM_NTASKS if it exists
When running in a SLURM session (either interactive or batch), we need
to make sure that OMP_NUM_THREADS is set to the same number of CPUs
that were requested in the SLURM environment.

We have now added an IF statement to .bashrc_OD that will set
OMP_NUM_THREADS to the SLURM_NTASKS variable, if it exists.  (This
variable only exists when running in an interactive or batch SLURM
environment).  Otherwise, OMP_NUM_THREADS will default to 1.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>