You can now specify the partition in interactive and interactive_openmp
Modified the bin/interactive and bin/interactive_openmp scripts to
allow you to specify the partition.  If the partition is not specified,
it will open an interactive SLURM session on the jacob partition (i.e.
on the Jacob-group dedicated nodes).

Also printed a reminder to have the user doublecheck the setting
of OMP_NUM_THREADS.  You could have requested 4 CPUs, for example,
but if OMP_NUM_THREADS is still set to 1, then OpenMP jobs will only
execute with 1 CPU.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>