Add bin/awake.sh script to keep Odyssey interactive sessions alive
The bin/awake.sh script can be run in the background of your Odyssey
interactive sessions.  The script will print out a message (currently
"...") to the screen every 10 minutes.  This should be sufficient to
prevent the interactive session from freezing at random times.

RC says they know about the freezing of the interactive sessions,
but it is not a priority for them to fix it at this time.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>