bin/jobstats now ignores extraneous memory values
We modified the bin/jobstats script to ignore extranous memory values
as reported by sacct.  Under some conditions, the jobinfo script (which
jobstats calls) can report the following:

  JobName          : HEMCO+MW.run          batch
  . . .
  MaxRSS           : 16?                   6.3307 GB

which causes jobinfo to report memory as such:

  Memory                : 16? 6.3307

We have now added a fix to jobstats so that it will ignore the
extraneous memory value (i.e. "16?") and will once again report
memory usage as:

  Memory                : 6.3307 GB

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>