Added bin/isCoards to test netCDF files for COARDS compliance
We have created a perl script (bin/isCoards) to test a netCDF file
for coards compliance.  This will help us to determine if netCDF files
(esp. those that we receive for HEMCO emissions) comply to the COARDS
set of standards.

The isCoards script will also be placed into the NcdfUtil folder
of the GEOS-Chem source code.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>