Bug fix: isCoards now properly eport non-monotonic index variable data
A minor bug was corrected in routine &isMonotonic of the bin/isCoards
script.  If non-monotonic data was encountered, the script formerly
returned -1 which was not correct.  It now returns a string
"is not monotonic", which is looked for in the various &analyze*
routines.

Also added print statements directing the user to look at a
certain page on the GC wiki for more information about how to use
NCO operators to fix items that are not COARDS-compliant.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>