Updated .bashrc_OD for PGI compiler modules etc.
(1) PGI compiler now loads openmpi v1.8.10 and netcdf v4.1.3.

(2) If using the GEOS-Chem-Libraries, we have to manually set the
    FC, CC, and CXX environment variables.

(3) No longer load totalview automatically; users can load that
    if they wish.

(4) If $NETCDF_HOME is blank, then set it to `nc-config --prefix`.

(5) If $NETCDF_INCLUDE is not blank, then use that to define $GC_INCLUDE;
    otherwise set it to "$NETCDF_HOME/include" by default.

(6) If $NETCDF_LIB is not blank, then use that to define $GC_LIB;
    otherwise set it to "$NETCDF_HOME/lib" by default.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>