Use OMP_STACKSIZE instead of KMP_STACKSIZE in .bashrc files
To make it easier to switch between Intel Fortran and GNU Fortran,
we now use the OMP_STACKSIZE variable to set the max amount of
stack memory.  (Actually, we request 500 MB of stack memory, which
should be a large enough value to max out the stack memory on all
systems.)

The KMP_STACKSIZE variable only works with Intel Fortran.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>