Update the jobinfo and jobstats script for new SLURM output format
We have now updated the jobinfo and jobstats scripts so that they can
better parse the SLURM output info via sacct.

In a recent SLURM update, the output of sacct can now be split amongst
columns labeled "batch" or "time".  We have reworked the scripts so
that they will always find the information from the proper columns.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>