Update sample run script to remove #SBATCH --mail-user option
If #SBATCH --mail-user is ommitted, SLURM will send an email to the
submitting user by default. This is more desirable than forcing users
to remember to change this option to their own email address.

Also removed the #SBATCH -o and #SBATCH -e options. When these are ommitted,
SLURM will send both stdout and stderr output to the slurm-JOBID.out file.
Again, this is more desirable than forcing users to change the file name
for the version that they are using.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>