Update module load commands in .bashrc_AS
Make sure the latest versions of modules are loaded on AS. An update to CDO
prompted this change. CDO v1.6.4 was not working properly on AS, so v1.7.2
was installed. To always load the latest available version of a module, we
shouldn't specify a version number in the module load command.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>