Allow user to manually specify cores used as an argument to jobstats
SLURM only reports the cores that were allocated for a job but not
the cores that were actually used.  This can cause the computation of
the "% ideal" quantity to be skewed for those runs where a whole node
is requested but only some of the cores on the node are used.

To correct this, the jobstats script now accepts the number of cores
as the second argument.  For example:

  jobstats 33939645     # assumes 32 cores, prints "(31.03% ideal)"
  jobstats 33939645 12  # uses 12 cores,    prints "(82.75% ideal)"

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>