Update .emacs file to always show the menu bar
FASRC has built a new version of Emacs (26.1) with a properly-
functioning tool bar.  We have modified the emacs file to
(1) Always show the toolbar by default and (2) show 2 less lines
of text to make room for the toolbar.

Users can of course modify the number of lines of text that emacs
starts with at the end of the .emacs file.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>