Turn off automatic indenting for F90 mode in root/.emacs
We have disabled the electric-indent-mode (which does automatic
indentation) whenever F90 mode is activated.

In versions of Emacs newer than 24.4, electric-indent-mode is always
turned on.  But this will indent all text and comment characters
to the current indentation level.  Therefore, if you wanted to put a
comment character in the first column, you would have to manually
backspace, which is annoying.

You can always use the TAB key to align code, as we already do
for Fortran (aka F77) mode.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>