Now pass optional argument for # of OpenMP threads to init scritps
We have modified the init/init*centos7 scripts so that they can
accept an optional argument, which can then be used to set the
value of $OMP_NUM_THREADS.

If you are in a SLURM interactive or batch session, then
OMP_NUM_THREADS will still be set to the value of
SLURM_CPUS_PER_TASK automatically.  But if you are not in a SLURM
session, then you can manually specify the number of OpenMP
threads to use.

This is more or less a "hidden feature".

Also removed init scripts for CentOS6.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>